Conflicts:
	Makefile
	Server/AudioManager.h
	Server/Client.cpp
	Server/Entity.cpp
	Server/graphics/drawSet.cpp
	Server/graphics/drawSet.h
	Server/graphics/graphicsClient.cpp

+++ Server/Projectile.cpp
@@ -3,7 +3,10 @@
 Projectile::Projectile(Map* m)
 {
 	this->map = m;
-
+  fired = false;
+  charge_level = FIREONE;
+  charging = true;
+  charge_counter = sf::Clock();
   BoundingBox* b = new BoundingBox(BoundingObj::vec4_t(0,0,0),BoundingObj::vec3_t(1,0,0),BoundingObj::vec3_t(0,1,0),BoundingObj::vec3_t(0,0,1),
   1,1,1);
   b->setEntity(this);
@@ -14,18 +17,38 @@ Projectile::~Projectile(void)
 {
 }
 
-void Projectile::update(void) {
-  clearEvents();
-  v3_t distanceTravelled = velocity * ConfigManager::serverTickLengthSec();
-	position += distanceTravelled;
-  
-  //see if travelled full range
-  distanceLeftToTravel -= distanceTravelled.magnitude();
-  if(distanceLeftToTravel <= 0.0){
-    map->destroyProjectile(this);
-    return;
+bool Projectile::correctMovementHit( Entity* e ){
+  Entity_Type etype = e->getType();
+  if( etype == WALL || etype == PROJECTILE ){
+    return true;
+  } else if ( etype == PLAYER ){
+    return e != owner;
   }
+  return false;
+}
 
+void Projectile::update(void) {
+
+    if(charging && charge_level != FIRETHREE ) {
+      if(charge_counter.getElapsedTime().asMilliseconds() > Charge_Time ) {
+        charge_counter.restart();
+        charge_level = (MAGIC_POWER)(charge_level  + 1);
+        std::cout << "increase level heeeeer !" << std::endl;
+      }
+      return;
+    }
+    v3_t distanceTravelled = velocity * ConfigManager::serverTickLengthSec();
+  distanceTravelled = correctMovement(distanceTravelled, false);
+	  position += distanceTravelled;
+  
+    //see if travelled full range
+    distanceLeftToTravel -= distanceTravelled.magnitude();
+    if(distanceLeftToTravel <= 0.0){
+      map->destroyProjectile(this);
+      return;
+    }
+
+  
 	updateBounds();
 	// some collision detection
 }
@@ -36,6 +59,10 @@ void Projectile::setOwner(Player * player){
 
 void Projectile::setStrength(float f) {
 	strength = f;
+}
+
+float Projectile::getStrength() {
+  return strength*magic_mutiplier[charge_level];
 }
 
 void Projectile::setRange(length_t r) {
@@ -59,15 +86,32 @@ void Projectile::handleCollisions() {
 
   for( auto it = entities.begin(); it != entities.end(); it++ ){
     Entity * e = it->first; 
-    if(e != owner) {
+    if(e != owner &&  e->getType() != PROJECTILE) {
        map->destroyProjectile(this);
     }
   }
-
 }
 
 void Projectile::clearEvents(){
-  if(!firedGuard)
-    fired = false;
-  firedGuard = false;
+  fired = false;
 }
+
+void Projectile::fire(v3_t v) {
+  velocity = v;
+  fired = true;
+  charging = false;
+}
+
+void Projectile::serialize(sf::Packet & packet) const {
+  Entity::serialize(packet);
+  packet << fired;
+  //(*owner).serialize(packet);
+}
+
+void Projectile::deserialize( sf::Packet & packet ) {
+  Entity::deserialize(packet);
+  packet >> fired;
+  //delete owner; this segfaults
+  //Player* owner = new Player();
+  //(*owner).deserialize(packet);
+ }
